Removed embedding layer from encoder
The model was not able to learn due to embedding layer cause when we feed the true
input to the model via the embedding layer and since the embedding layer is also
learning, the true input did not reach the model properly. This made the network VERY
VERY unstable. So, I removed the embedding layer and added one hot vectores to encode
the y_in. And the model, starts to learn. Woohooo!! Happy new year.